Added TestBuilds
The purpose of TestBuilds is to ensure that an exported jar file will run properly. This will require the most recent copy of the assets folder to be placed in the same folder as the jar file. Please date any entries in the test builds folder in a separate folder in the mmddyyyy format, and format your jar files in the mmddyyyyv* format, where * is the number of times you've exported a jar file to the folder in a given day

Updated the README file to reflect this specification